requirements.txt
Django==2.1.4
djangorestframework==3.9.0
mixer==6.1.3
django-filter==2.0.0